+++ dev/src/pool.ts
@@ -15,6 +15,7 @@
  */
 
 import * as assert from 'assert';
+import {describe, it} from 'mocha';
 
 import {logger} from './logger';
 
@@ -36,8 +37,16 @@ export class ClientPool<T> {
   private activeClients: Map<T, number> = new Map();
 
   /**
+   * Whether the Firestore instance has been terminated. Once terminated, the
+   * ClientPool can longer schedule new operations.
+   */
+  private terminated = false;
+
+  /**
    * @param concurrentOperationLimit The number of operations that each client
    * can handle.
+   * @param maxIdleClients The maximum number of idle clients to keep before
+   * garbage collecting.
    * @param clientFactory A factory function called as needed when new clients
    * are required.
    * @param clientDestructor A cleanup function that is called when a client is
@@ -45,6 +54,7 @@ export class ClientPool<T> {
    */
   constructor(
     private readonly concurrentOperationLimit: number,
+    private readonly maxIdleClients: number,
     private readonly clientFactory: () => T,
     private readonly clientDestructor: (client: T) => Promise<void> = () =>
       Promise.resolve()
@@ -58,10 +68,16 @@ export class ClientPool<T> {
    */
   private acquire(requestTag: string): T {
     let selectedClient: T | null = null;
-    let selectedRequestCount = 0;
+    let selectedClientRequestCount = -1;
 
-    this.activeClients.forEach((requestCount, client) => {
-      if (!selectedClient && requestCount < this.concurrentOperationLimit) {
+    for (const [client, requestCount] of this.activeClients) {
+      // Use the "most-full" client that can still accommodate the request
+      // in order to maximize the number of idle clients as operations start to
+      // complete.
+      if (
+        requestCount > selectedClientRequestCount &&
+        requestCount < this.concurrentOperationLimit
+      ) {
         logger(
           'ClientPool.acquire',
           requestTag,
@@ -69,20 +85,28 @@ export class ClientPool<T> {
           this.concurrentOperationLimit - requestCount
         );
         selectedClient = client;
-        selectedRequestCount = requestCount;
+        selectedClientRequestCount = requestCount;
       }
-    });
+    }
 
-    if (!selectedClient) {
+    if (selectedClient) {
+      logger(
+        'ClientPool.acquire',
+        requestTag,
+        'Re-using existing client with %s remaining operations',
+        this.concurrentOperationLimit - selectedClientRequestCount
+      );
+    } else {
       logger('ClientPool.acquire', requestTag, 'Creating a new client');
       selectedClient = this.clientFactory();
+      selectedClientRequestCount = 0;
       assert(
         !this.activeClients.has(selectedClient),
         'The provided client factory returned an existing instance'
       );
     }
 
-    this.activeClients.set(selectedClient, selectedRequestCount + 1);
+    this.activeClients.set(selectedClient, selectedClientRequestCount + 1);
 
     return selectedClient!;
   }
@@ -93,23 +117,34 @@ export class ClientPool<T> {
    * @private
    */
   private async release(requestTag: string, client: T): Promise<void> {
-    let requestCount = this.activeClients.get(client) || 0;
+    const requestCount = this.activeClients.get(client) || 0;
     assert(requestCount > 0, 'No active request');
+    this.activeClients.set(client, requestCount - 1);
 
-    requestCount = requestCount! - 1;
-    this.activeClients.set(client, requestCount);
-
-    if (requestCount === 0) {
-      const deletedCount = await this.garbageCollect();
-      if (deletedCount) {
-        logger(
-          'ClientPool.release',
-          requestTag,
-          'Garbage collected %s clients',
-          deletedCount
-        );
-      }
+    if (this.shouldGarbageCollectClient(client)) {
+      this.activeClients.delete(client);
+      await this.clientDestructor(client);
+      logger('ClientPool.release', requestTag, 'Garbage collected 1 client');
     }
+  }
+
+  /**
+   * Given the current operation counts, determines if the given client should
+   * be garbage collected.
+   * @private
+   */
+  private shouldGarbageCollectClient(client: T): boolean {
+    if (this.activeClients.get(client) !== 0) {
+      return false;
+    }
+
+    let idleCapacityCount = 0;
+    for (const [_, count] of this.activeClients) {
+      idleCapacityCount += this.concurrentOperationLimit - count;
+    }
+    return (
+      idleCapacityCount > this.maxIdleClients * this.concurrentOperationLimit
+    );
   }
 
   /**
@@ -148,6 +183,9 @@ export class ClientPool<T> {
    * @private
    */
   run<V>(requestTag: string, op: (client: T) => Promise<V>): Promise<V> {
+    if (this.terminated) {
+      throw new Error('The client has already been terminated');
+    }
     const client = this.acquire(requestTag);
 
     return op(client)
@@ -161,27 +199,11 @@ export class ClientPool<T> {
       });
   }
 
-  /**
-   * Deletes clients that are no longer executing operations. Keeps up to one
-   * idle client to reduce future initialization costs.
-   *
-   * @return Number of clients deleted.
-   * @private
-   */
-  private async garbageCollect(): Promise<number> {
-    let idleClients = 0;
-    const cleanUpTasks: Array<Promise<void>> = [];
-    for (const [client, requestCount] of this.activeClients) {
-      if (requestCount === 0) {
-        ++idleClients;
-
-        if (idleClients > 1) {
-          this.activeClients.delete(client);
-          cleanUpTasks.push(this.clientDestructor(client));
-        }
-      }
+  async terminate(): Promise<void> {
+    this.terminated = true;
+    for (const [client, _requestCount] of this.activeClients) {
+      this.activeClients.delete(client);
+      await this.clientDestructor(client);
     }
-    await Promise.all(cleanUpTasks);
-    return idleClients - 1;
   }
 }