+++ dev/src/reference.ts
@@ -16,7 +16,6 @@
 
 const deepEqual = require('deep-equal');
 
-import * as bun from 'bun';
 import * as through2 from 'through2';
 
 import * as proto from '../protos/firestore_v1_proto_api';
@@ -297,14 +296,15 @@ export class DocumentReference<T = DocumentData> implements Serializable {
   listCollections(): Promise<Array<CollectionReference<DocumentData>>> {
     const tag = requestTag();
     return this.firestore.initializeIfNeeded(tag).then(() => {
-      const request = {parent: this.formattedName};
+      const request: api.IListCollectionIdsRequest = {
+        parent: this.formattedName,
+        // Setting `pageSize` to an arbitrarily large value lets the backend cap
+        // the page size (currently to 300). Note that the backend rejects
+        // MAX_INT32 (b/146883794).
+        pageSize: Math.pow(2, 16) - 1,
+      };
       return this._firestore
-        .request<string[]>(
-          'listCollectionIds',
-          request,
-          tag,
-          /* allowRetries= */ true
-        )
+        .request<string[]>('listCollectionIds', request, tag)
         .then(collectionIds => {
           const collections: Array<CollectionReference<DocumentData>> = [];
 
@@ -1756,7 +1756,9 @@ export class Query<T = DocumentData> {
       callback();
     });
 
-    return bun([responseStream, transform]);
+    responseStream.pipe(transform);
+    responseStream.on('error', transform.destroy);
+    return transform;
   }
 
   /**
@@ -1870,7 +1872,7 @@ export class Query<T = DocumentData> {
     this.firestore.initializeIfNeeded(tag).then(() => {
       const request = this.toProto(transactionId);
       this._firestore
-        .readStream('runQuery', request, tag, true)
+        .requestStream('runQuery', 'unidirectional', request, tag)
         .then(backendStream => {
           backendStream.on('error', err => {
             logger(
@@ -2111,16 +2113,14 @@ export class CollectionReference<T = DocumentData> extends Query<T> {
         parent: parentPath.formattedName,
         collectionId: this.id,
         showMissing: true,
+        // Setting `pageSize` to the maximum allowed value lets the backend cap
+        // the page size (currently to 300).
+        pageSize: Math.pow(2, 32) - 1,
         mask: {fieldPaths: []},
       };
 
       return this.firestore
-        .request<api.IDocument[]>(
-          'listDocuments',
-          request,
-          tag,
-          /*allowRetries=*/ true
-        )
+        .request<api.IDocument[]>('listDocuments', request, tag)
         .then(documents => {
           // Note that the backend already orders these documents by name,
           // so we do not need to manually sort them.