+++ dev/test/collection.ts
@@ -149,6 +149,7 @@ describe('Collection interface', () => {
           parent: `${DATABASE_ROOT}/documents/a/b`,
           collectionId: 'c',
           showMissing: true,
+          pageSize: 4294967295,
           mask: {fieldPaths: []},
         });