+++ dev/test/document.ts
@@ -46,7 +46,7 @@ import {
   writeResult,
 } from './util/helpers';
 
-import api = proto.google.firestore.v1;
+import {GoogleError, Status} from 'google-gax';
 
 const PROJECT_ID = 'test-project';
 
@@ -142,7 +142,7 @@ describe('serialize document', () => {
     expect(() => {
       firestore.doc('collectionId/documentId').set({foo: undefined});
     }).to.throw(
-      'Value for argument "data" is not a valid Firestore document. Cannot use "undefined" as a Firestore value (found in field foo).'
+      'Value for argument "data" is not a valid Firestore document. Cannot use "undefined" as a Firestore value (found in field "foo").'
     );
 
     expect(() => {
@@ -150,14 +150,14 @@ describe('serialize document', () => {
         foo: FieldPath.documentId(),
       });
     }).to.throw(
-      'Value for argument "data" is not a valid Firestore document. Cannot use object of type "FieldPath" as a Firestore value (found in field foo).'
+      'Value for argument "data" is not a valid Firestore document. Cannot use object of type "FieldPath" as a Firestore value (found in field "foo").'
     );
 
     expect(() => {
       class Foo {}
       firestore.doc('collectionId/documentId').set({foo: new Foo()});
     }).to.throw(
-      'Value for argument "data" is not a valid Firestore document. Couldn\'t serialize object of type "Foo" (found in field foo). Firestore doesn\'t support JavaScript objects with custom prototypes (i.e. objects that were created via the "new" operator).'
+      'Value for argument "data" is not a valid Firestore document. Couldn\'t serialize object of type "Foo" (found in field "foo"). Firestore doesn\'t support JavaScript objects with custom prototypes (i.e. objects that were created via the "new" operator).'
     );
 
     expect(() => {
@@ -412,30 +412,6 @@ describe('deserialize document', () => {
     });
   });
 
-  it('ignores intermittent stream failures', () => {
-    let attempts = 1;
-
-    const overrides: ApiOverride = {
-      batchGetDocuments: () => {
-        if (attempts < 3) {
-          ++attempts;
-          throw new Error('Expected error');
-        } else {
-          return stream(found(document('documentId')));
-        }
-      },
-    };
-
-    return createInstance(overrides).then(firestore => {
-      return firestore
-        .doc('collectionId/documentId')
-        .get()
-        .then(() => {
-          expect(attempts).to.equal(3);
-        });
-    });
-  });
-
   it('deserializes date before 1970', () => {
     const overrides: ApiOverride = {
       batchGetDocuments: () => {
@@ -676,7 +652,9 @@ describe('get document', () => {
   it('throws error', done => {
     const overrides: ApiOverride = {
       batchGetDocuments: () => {
-        return stream(new Error('RPC Error'));
+        const error = new GoogleError('RPC Error');
+        error.code = Status.PERMISSION_DENIED;
+        return stream(error);
       },
     };
 
@@ -1262,7 +1240,7 @@ describe('set document', () => {
     expect(() => {
       firestore.doc('collectionId/documentId').set({foo: FieldValue.delete()});
     }).to.throw(
-      'Value for argument "data" is not a valid Firestore document. FieldValue.delete() must appear at the top-level and can only be used in update() or set() with {merge:true} (found in field foo).'
+      'Value for argument "data" is not a valid Firestore document. FieldValue.delete() must appear at the top-level and can only be used in update() or set() with {merge:true} (found in field "foo").'
     );
   });
 
@@ -1611,7 +1589,7 @@ describe('update document', () => {
         a: {b: FieldValue.delete()},
       });
     }).to.throw(
-      'Update() requires either a single JavaScript object or an alternating list of field/value pairs that can be followed by an optional precondition. Value for argument "dataOrField" is not a valid Firestore value. FieldValue.delete() must appear at the top-level and can only be used in update() or set() with {merge:true} (found in field a.b).'
+      'Update() requires either a single JavaScript object or an alternating list of field/value pairs that can be followed by an optional precondition. Value for argument "dataOrField" is not a valid Firestore value. FieldValue.delete() must appear at the top-level and can only be used in update() or set() with {merge:true} (found in field "a.b").'
     );
 
     expect(() => {
@@ -1619,7 +1597,7 @@ describe('update document', () => {
         b: FieldValue.delete(),
       });
     }).to.throw(
-      'Update() requires either a single JavaScript object or an alternating list of field/value pairs that can be followed by an optional precondition. Element at index 1 is not a valid Firestore value. FieldValue.delete() must appear at the top-level and can only be used in update() or set() with {merge:true} (found in field a.b).'
+      'Update() requires either a single JavaScript object or an alternating list of field/value pairs that can be followed by an optional precondition. Element at index 1 is not a valid Firestore value. FieldValue.delete() must appear at the top-level and can only be used in update() or set() with {merge:true} (found in field "a.b").'
     );
 
     expect(() => {
@@ -2005,6 +1983,7 @@ describe('listCollections() method', () => {
       listCollectionIds: (request, options, callback) => {
         expect(request).to.deep.eq({
           parent: `projects/${PROJECT_ID}/databases/(default)/documents/coll/doc`,
+          pageSize: 65535,
         });
 
         callback(null, ['second', 'first']);